Modernization/refactoring.
Add override flags to functions in Camera.
Replace for loop with range based in Frame.cpp.
Correctly move the string in the VideoStitcher constructor.
For->range-based in View.cpp
Correct use of dereference function in declaring autos.
Change 0 to nullptr in MatView and RealtimeViewer.
Correct old for-loop in main.cpp.